fix bugs and add select for update in db

+++ app/wallet/schemas.py
@@ -2,9 +2,9 @@ from datetime import datetime
 from typing import Literal
 from pydantic import BaseModel
 from pydantic import UUID4
-
+from decimal import Decimal
 
 
 class OperationRequest(BaseModel):
     operationType: Literal["DEPOSIT", "WITHDRAW"]
-    amount: float
+    amount: Decimal
